Move PlaceHolder iterator class
